requirements.txt
aniso8601==8.0.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
idna==3.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.14.0
marshmallow-sqlalchemy==0.26.1
passlib==1.7.4
psycopg2==2.9.2
pycodestyle==2.8.0
PyJWT==2.3.0
python-dotenv==0.19.1
pytz==2020.1
requests==2.26.0
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.2
urllib3==1.26.7
Werkzeug==1.0.1